Email_pipeline/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
python-dotenv>=1.0.0
pypdf>=4.0.0
pandas>=2.2.0
openai>=1.40.0
